requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
scikit-surprise>=1.1.3
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
